fix: Execute repositories before packages
If a package to be installed is provided by a repository to be added, it
will cause the first apply to fail, because packages state is executed
before repositories state. Next apply will succeed, as the repository
was setup in the previous one.

Changed state execution order so repositories is executed before
packages.