juniper_junos: refine default interface MTU
In simulated (vSRX/vQFX) environments the big MTU exceeded some of the
interface maximums. As testing should not be affected by skipping jumbo
packets, a "simulation" pillar entry is introduced which can be
dynamically passed with Salt commands in the test environment. If it is
set to True, the default MTU will be lowered to 1500.

Signed-off-by: Georg Pfuetzenreuter <georg.pfuetzenreuter@suse.com>