Repair backup handling in saltsafe_if
In some situations, the script might fail with an unbound variable due
to backup() being called before the $file_backup declaration. This is
now resolved - additionally, a check for an existing origin file is
implemented, in case backup() is called before the origin file has been
written for the first time.

Signed-off-by: Georg Pfuetzenreuter <georg.pfuetzenreuter@suse.com>