kexec: repair conditional execution
The "kexec -S" call checks "kexec_crash_loaded", which is only
relevant for Kdump. In this case we are interested in "kexec_loaded".
Additionally, replace the service state with a service module call,
since the service.running state function tends to fail with oneshot
services, not being able to detect them alive after execution.

Signed-off-by: Georg Pfuetzenreuter <georg.pfuetzenreuter@suse.com>