rebootmgr: repair configuration for Tumbleweed
On Tumbleweed systems the configuration file placed using the formula
code was missing the "[rebootmgr]" header, causing options to be ignored.
This issue only affects Tumbleweed, where no /etc/rebootmgr.conf file
is provided by the package in favor of the one in /usr/etc. On Leap,
the package provides the file under /etc including the necessary section
header.
To resolve this problemtic discrepancy as well as to simplify the logic,
this refactors the handling to a simple file.managed state containing both
the common and the section header along with the configuration options.
This will ensure a uniform and functional file across both Leap and
Tumbleweed systems.

Signed-off-by: Georg Pfuetzenreuter <georg.pfuetzenreuter@suse.com>