rebootmgr: repair configuration reset on Leap
Upon unconfiguring the formula (i.e. removing the "rebootmgr" pillar
or no longer including it on a minion), the configuration should
reset to the defaults.
On Tumbleweed this happens by removing the administrative configuration
file and having the software fall back to the package defaults.
This is not possible on Leap due to packaging constraints, hence
revert the administrative configuration to the formula defaults
(which aim to reflect the package defaults) as a substitute.

Signed-off-by: Georg Pfuetzenreuter <mail@georg-pfuetzenreuter.net>