Repair error on initial network.wicked execution
The interface states rely on a helper script - if the common
states have not yet been applied on a minion, an initial test run
would return network_wicked_interfaces as failed due to the helper
script not being present.
Resolve this by testing whether the script has already been installed,
and constructing a graceful return otherwise.

Signed-off-by: Georg Pfuetzenreuter <mail@georg-pfuetzenreuter.net>