apache_httpd: avoid conflicting listeners
The server cannot bind to a port on a specific address if the
same port is already bound to a wildcard listener. Handle this
by detecting and removing listeners bound to a specific address
if the same port is already configured in a wildcard listener.

This helps with constructing pillar based vhost overrides with
different listener configurations and makes usage of the formula
less prone to errors.

Signed-off-by: Georg Pfuetzenreuter <mail@georg-pfuetzenreuter.net>