apache_httpd: repetitive options as mappings
This adds suport for mappings which translate to multiple lines
instead of "< ... > ... </ ... >" blocks, extending the existing
"repetitive options" logic.
As an initial configuration option to utilize this, "SetEnvIf"
is added to the repetitive options.

Previously, it was needed to provide "SetEnvIf" configuration in
the following format:

```
  configs:
    example:
      SetEnvIf:
        Request_URI: ^/example$ myvariable
```

With this patch, the following will render the same configuration:

```
  configs:
    example:
      SetEnvIf:
        Request_URI:
          ^/example$: myvariable
```

This allows for a more logical data structure in the pillar,
uniform with what is already used for less nested options, making
the formula easier to parse and less error prone to use.

Signed-off-by: Georg Pfuetzenreuter <mail@georg-pfuetzenreuter.net>