apache_httpd: query pillar inside templates
The "context" approach fails when patterns requiring special escaping,
most prominently regular expressions, are attempted to be passed
through the pillar, because of the escaping only persisting through
the first YAML abstraction (the pillar) and not through the second
one (the state file).
To avoid this pitfall, skip the second YAML layer by retrieving the
pillar directly in the Jinja templates instead of passing it
through "context".

Signed-off-by: Georg Pfuetzenreuter <mail@georg-pfuetzenreuter.net>