infrastructure.salt: add qjid script
For querying jobs with lots of output it is often useful to
utilize a pager, but one does not want to miss out on color when
doing so. Combining these desires makes for a rather long command
to type, and reusing it requires the need to replace the ID
inconveniently located in the middle.
Make this easier by adding a small helper script.

Signed-off-by: Georg Pfuetzenreuter <mail@georg-pfuetzenreuter.net>