infrastructure: Python pillar library
This imports the Python based pillar logic used in our Salt
infrastructures in the form of reusable libraries, helping us to
deduplicate code and making it easier to maintain changes.
The respective "#!py" pillars will then be adjusted to reference library
calls, with only the infrastructure specific data needing to be passed
as function parameters.

Signed-off-by: Georg Pfuetzenreuter <mail@georg-pfuetzenreuter.net>