infra.suse_ha: block on failed migrations
By default a migration failure, for example due to an unreachable
Libvirt TCP socket on the target hypervisor, would cause the virtual
machine resource to get stopped on the source, and then started again on
the target hypervisor. In our infrastructure this behavior is not
desirable - rather abort the operation and leave the VM running on the
source hypervisor, allowing an administrator to resolve the migration issue
without having to worry about interrupted services due to unplanned VM
restarts.

Signed-off-by: Georg Pfuetzenreuter <mail@georg-pfuetzenreuter.net>