Update year and description in spec template
The summary and description make it sound like as if the states were
specific to our infrastructure, but this is only the case for the
infrastructure-formula subpackage, which already reflects this in its
summary and description. Hence make the main package more welcoming and
update the year whilst at it.

Signed-off-by: Georg Pfuetzenreuter <mail@georg-pfuetzenreuter.net>