Preserve hyphen in packaged state directories
Avoid installing a state directory with an underscore if it is present
with a hyphen in the repository to allow for references in the templates
(for example to a map file) to resolve and to reduce confusion, as one
generally expects the state directory name in the repository to match
the installed one.
This makes the metadata directory diverge, as that one is still installed
by the top formula name, however the metadata matching the formula name is
deemed expected.

This is a breaking change, the packaged state "os_update" changes to
"os-update" in the next spec rendering.
The packaged "php_fpm" state changing to "php-fpm" is not breaking, as it
was broken already and is hereby repaired.

Signed-off-by: Georg Pfuetzenreuter <mail@georg-pfuetzenreuter.net>