infrastructure.salt: handle existing GPG keys
If the generated key is not the first one in the keyring, the query
would yield a sequence of fingerprints concatenated by newlines, causing
the subsequent commands to fail due to not being passed a valid (single)
fingerprint.
Repair this by only querying for the generated key - this assumes that
no existing keys use the same email address, which is deemed reliable
enough as the script already asserts whether the passed email address
is already in use.

Signed-off-by: Georg Pfuetzenreuter <georg.pfuetzenreuter@suse.com>