infrastructure.salt: set syndic key permissions
The minion and syndic share a keypair.
The private key file gets created with 0400 permissions but owned by the
user running the minion process - which is root - causing failure of the
syndic service which cannot access the key file due to running under a less
privileged user.
Enforce the existing restrictive file mode but together with access to
the salt user running the syndic process.

Signed-off-by: Georg Pfuetzenreuter <mail@georg-pfuetzenreuter.net>