network: support systemd.link management
Initial handling to allow for preparing interfaces then referenced in
other network states. The directory is managed by the udev package,
which should already be installed on managed systems through other
means. The files are in ini format, but the custom template was deemed
more efficient than the ini management states, and avoids the need for
additional states to manage the header.

Signed-off-by: Georg Pfuetzenreuter <mail@georg-pfuetzenreuter.net>