network.wicked: improve clean default
The main formula state, which includes interfaces including the cleanup
section, is also getting included on minions which so far had their
/etc/sysconfig/network/config but not their interfaces managed through
the formula. The new default caused undesired deletions. Improve the
behavior if no explicit "clean" choice is defined: if interfaces are
declared in the pillar, it can be assumed that all interfaces on the
machine should be managed and enforced, but if no interfaces are
declared in the pillar, it can be assumed that the machine is operating
with intentionally manual ones.

Signed-off-by: Georg Pfuetzenreuter <mail@georg-pfuetzenreuter.net>